기본 데이터 단위 변경

+++ Assets/04 Items/Data/ItemData.cs
@@ -41,6 +41,6 @@ public class ItemData : ScriptableObject
     [Header("Equip Stats")]
     public int attackPower;
     public int shieldPower;
-    public float moveSpeedBonus;
+    public int moveSpeedBonus;
     public int maxHpBonus;
 }
